Fix Faker::Date.between_except
Faker::Date.between_except causes an infinite loop when the same date is
passed as all the three parameters.

The `begin ... end` construct has been replaced by `loop ... do`, since
the use of the former has been discouraged by Matz himself.

Also, the excepted date is not converted to date object when it is
passed as a string, and this leads to a comparison between string and
date which causes the excepted date being actually included in the
results.

This commit fixes the wrong behaviour and raise an ArgumentError when
the method is called with incompatible parameters.

Ref: http://blade.nagaokaut.ac.jp/cgi-bin/scat.rb/ruby/ruby-core/6741