Fix the path names for bg.yml (#1877)
The fields doesn't need to use the 'Address' prefix for referring as
it's already inside the `address` subfield for this locale. Note that
other locales like `ar.yml`, follow this properly.

Note that having it like this affects my Haskell client which relies
on the faker's data source. Relevant MR which found this bug:
https://github.com/psibi/fakedata/pull/9