Various locale specific and other fixes for faker
Summary of the changes:

- touhou.yml: The current file isn't a valid yaml spec. Use single
  quote to fix it.
- es-AR: Remove redundant entires, the `country_code` should be an
array (as consistent with the other formats)
- fr: Break `appliance.yml` from `pokemon.yml` as it's a separate
one. (Also makes it consistent with the en locale)
- ja: Break `animal.yml`, `cat.yml`, `dog.yml`, `overwatch.yml`,
`pokemon.yml`, `super_mario.yml`, `zelda.yml` from `creature.yml` and
`games.yml`. This makes it consistent with the `en` locale too.
- en/music.yml: Move hiphop under music as that's the top level
attribute. This also makes it consistent with the other yml convention
we follow.