Freeze new generator and locales (#2886)
* Freeze new generator and locales

As we discussed here https://github.com/orgs/faker-ruby/discussions/2877
we want to improve faker's performance and organization. There hasn't been
to many bug reports and most of the open issues are related to performance
and confusion around using faker.

As we have limited time to invest in faker, reviewing new generators and
locales take lots of our time. With this decision, we hope to make the
necessary changes for faker to go to the next level.

* Reinforce what contributors we will accept during the freeze